+++ poko-gradle-plugin/src/main/kotlin/dev/drewhamilton/poko/gradle/PokoGradlePlugin.kt
@@ -1,6 +1,7 @@
 package dev.drewhamilton.poko.gradle
 
 import com.android.build.api.dsl.CommonExtension
+import dev.drewhamilton.poko.gradle.BuildConfig.DEFAULT_POKO_FIR_IDE_MODE
 import dev.drewhamilton.poko.gradle.BuildConfig.DEFAULT_POKO_ANNOTATION
 import dev.drewhamilton.poko.gradle.BuildConfig.DEFAULT_POKO_ENABLED
 import org.gradle.api.Project
@@ -19,6 +20,7 @@ public class PokoGradlePlugin : KotlinCompilerPluginSupportPlugin {
         val extension = target.extensions.create("poko", PokoPluginExtension::class.java)
         extension.enabled.convention(DEFAULT_POKO_ENABLED)
         extension.pokoAnnotation.convention(DEFAULT_POKO_ANNOTATION)
+        extension.firIdeMode.convention(PokoFirIdeMode.valueOf(DEFAULT_POKO_FIR_IDE_MODE))
 
         val pokoAnnotationDependency = extension.pokoAnnotation.map {
             when (it) {
@@ -85,6 +87,14 @@ public class PokoGradlePlugin : KotlinCompilerPluginSupportPlugin {
                 SubpluginOption(
                     key = BuildConfig.POKO_ANNOTATION_OPTION_NAME,
                     value = it,
+                )
+            }
+        )
+        optionsProvider.add(
+            extension.firIdeMode.map {
+                SubpluginOption(
+                    key = BuildConfig.POKO_FIR_IDE_MODE_OPTION_NAME,
+                    value = it.name,
                 )
             }
         )